+++ app/src/main/java/com/example/mgckiosk/menu/bakery/sub/Bread.kt
@@ -1,16 +1,17 @@
-package com.example.mgckiosk.menu.level3
+package com.example.mgckiosk.menu.bakery.sub
 
 import com.example.mgckiosk.abs.AbstractMenu
 import com.example.mgckiosk.action.NumberOfItems
 import com.example.mgckiosk.exception.IllegalArgumentException
 
-class StickCoffee(override var category: Int) : AbstractMenu() {
+class Bread(override var category: Int) : AbstractMenu() {
+
     override fun displayInfo() {
-        var info = stickCoffee
+        var info = bread
         if (category == -1) {
             category = 0
         } else {
-            println("[ 스틱 커피 상세 메뉴 ]")
+            println("[ 빵 상세 메뉴 ]")
             println("")
 
             for (i in info.indices) {
