+++ app/src/main/java/com/example/mgckiosk/menu/bakery/sub/CookieMacaron.kt
@@ -1,16 +1,16 @@
-package com.example.mgckiosk.menu.level3
+package com.example.mgckiosk.menu.bakery.sub
 
 import com.example.mgckiosk.abs.AbstractMenu
 import com.example.mgckiosk.action.NumberOfItems
 import com.example.mgckiosk.exception.IllegalArgumentException
 
-class IcedTea(override var category: Int) : AbstractMenu() {
+class CookieMacaron(override var category: Int) : AbstractMenu() {
     override fun displayInfo() {
-        var info = icedTea
+        var info = cookieMacaron
         if (category == -1) {
             category = 0
         } else {
-            println("[ 차가운 티 상세 메뉴 ]")
+            println("[ 쿠키 & 마카롱 상세 메뉴 ]")
             println("")
 
             for (i in info.indices) {
@@ -23,4 +23,5 @@ class IcedTea(override var category: Int) : AbstractMenu() {
             NumberOfItems().numberOfItems(this, info, category)
         }
     }
+
 }
