+++ app/src/main/java/com/example/mgckiosk/menu/beverage/sub/IcedTea.kt
@@ -1,16 +1,16 @@
-package com.example.mgckiosk.menu.level3
+package com.example.mgckiosk.menu.beverage.sub
 
 import com.example.mgckiosk.abs.AbstractMenu
 import com.example.mgckiosk.action.NumberOfItems
 import com.example.mgckiosk.exception.IllegalArgumentException
 
-class HotCoffee(override var category: Int) : AbstractMenu() {
+class IcedTea(override var category: Int) : AbstractMenu() {
     override fun displayInfo() {
-        var info = hotCoffee
+        var info = icedTea
         if (category == -1) {
             category = 0
         } else {
-            println("[ 따뜻한 커피 상세 메뉴 ]")
+            println("[ 차가운 티 상세 메뉴 ]")
             println("")
 
             for (i in info.indices) {
