+++ app/src/main/java/com/example/mgckiosk/menu/product/sub/TeaProduct.kt
@@ -1,16 +1,16 @@
-package com.example.mgckiosk.menu.level3
+package com.example.mgckiosk.menu.product.sub
 
 import com.example.mgckiosk.abs.AbstractMenu
 import com.example.mgckiosk.action.NumberOfItems
 import com.example.mgckiosk.exception.IllegalArgumentException
 
-class SmoothieFrappe(override var category: Int) : AbstractMenu() {
+class TeaProduct(override var category: Int) : AbstractMenu() {
     override fun displayInfo() {
-        var info = smoothieFrappe
+        var info = teaProduct
         if (category == -1) {
             category = 0
         } else {
-            println("[ 스무디 & 프라페 상세 메뉴 ]")
+            println("[ 티 상품 상세 메뉴 ]")
             println("")
 
             for (i in info.indices) {
